+++ src/main/kotlin/mpp/kotlin/backend/controller/EmployeeController.kt
@@ -27,7 +27,7 @@ class EmployeeController {
     fun login(@RequestBody loginRequest: LoginRequest): Map<String, Any> {
         val employee = employeeService.login(loginRequest.email, loginRequest.password)
         return if (employee != null) {
-            val token = tokenProvider.generateToken(loginRequest.email)
+            val token = tokenProvider.generateToken(loginRequest.email, "employee")
             mapOf("token" to token, "employee" to employee)
         } else {
             throw UnauthorizedException("Invalid data")
@@ -40,6 +40,7 @@ class EmployeeController {
         @RequestParam("count") count: Int,
         @RequestHeader("Authorization") token: String
     ): List<Employee> {
+        println(tokenProvider.getRoleFromToken(token))
         if (tokenProvider.validateToken(token)) {
             val all = employeeService.findAll().toList()
             val endIndex = (start + count).coerceAtMost(all.size)
@@ -49,7 +50,8 @@ class EmployeeController {
 
     @GetMapping("/{id}")
     fun findById(@PathVariable id: Int, @RequestHeader("Authorization") token: String): Employee {
-        if (tokenProvider.validateToken(token)) {
+        println(tokenProvider.getRoleFromToken(token))
+        if (tokenProvider.validateToken(token) && tokenProvider.getRoleFromToken(token) == "admin") {
             try {
                 return employeeService.findById(id)
             } catch (e: RuntimeException) {
@@ -62,7 +64,8 @@ class EmployeeController {
 
     @PostMapping()
     fun save(@RequestBody request: EmployeeRequest, @RequestHeader("Authorization") token: String) {
-        if (tokenProvider.validateToken(token)) {
+        println(tokenProvider.getRoleFromToken(token))
+        if (tokenProvider.validateToken(token) && tokenProvider.getRoleFromToken(token) == "admin") {
             val address = request.address
             val id = addressService.findOne(address)
             var employee = Employee(
@@ -87,7 +90,8 @@ class EmployeeController {
     fun update(
         @PathVariable id: Int, @RequestBody request: EmployeeRequest, @RequestHeader("Authorization") token: String
     ) {
-        if (tokenProvider.validateToken(token)) {
+        println(tokenProvider.getRoleFromToken(token))
+        if (tokenProvider.validateToken(token) && tokenProvider.getRoleFromToken(token) == "admin") {
             val address = request.address
             val idAddress = addressService.findOne(address)
             var employee = Employee(
@@ -111,7 +115,8 @@ class EmployeeController {
 
     @DeleteMapping("/{id}")
     fun delete(@PathVariable id: Int, @RequestHeader("Authorization") token: String) {
-        if (tokenProvider.validateToken(token)) {
+        println(tokenProvider.getRoleFromToken(token))
+        if (tokenProvider.validateToken(token) && tokenProvider.getRoleFromToken(token) == "admin") {
             try {
                 employeeService.delete(id)
             } catch (e: RuntimeException) {
